[feat] Button refactoring
