+++ components/InfoCard/InfoCard.tsx
@@ -10,7 +10,7 @@ import StarOn from "../../assets/icons/icon-star-on.png";
 import BookmarkDisable from "../../assets/icons/icon-bookmark-disable.png";
 import BookmarkOff from "../../assets/icons/icon-bookmark-off.png";
 import BookmarkOn from "../../assets/icons/icon-bookmark-on.png";
-import CustomButton from "../Button/CustomButton";
+import CustomButton from "../Button/Button";
 
 export interface InfoCardProps {
   image?: string;
