+++ screens/ProfileUpdateScreen/ProfileUpdateScreen.tsx
@@ -3,7 +3,7 @@ import { View, Button, ScrollView, Image } from 'react-native';
 
 // image
 import ImageDefault from "../../assets/images/image-default.png";
-import CustomButton from '../../components/Button/CustomButton';
+import CustomButton from '../../components/Button/Button';
 import Input from '../../components/Input/Input';
 
 const ProfileUpdateScreen = () => {
