+++ screens/ReviewScreen/components/UseHint/UseHint.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { View, Text, StyleSheet } from "react-native";
 import "../../../../styles/colors";
 import CustomCheckBox from "../../../../components/CheckBox/CustomCheckBox";
-import CustomButton from "../../../../components/Button/CustomButton";
+import CustomButton from "../../../../components/Button/Button";
 
 const UseHint = () => {
   const [useHint, setUseHint] = useState(0);
@@ -24,7 +24,7 @@ const UseHint = () => {
         border="square"
         value="1개"
         onPress={() => setUseHint(1)} // 버튼을 누르면 힌트 사용 횟수 설정
-        // disabled={!isEnabled} // isEnabled가 false이면 버튼 비활성화
+      // disabled={!isEnabled} // isEnabled가 false이면 버튼 비활성화
       />
       <CustomButton
         size="small"
@@ -32,7 +32,7 @@ const UseHint = () => {
         border="square"
         value="2개"
         onPress={() => setUseHint(2)}
-        // disabled={!isEnabled}
+      // disabled={!isEnabled}
       />
       <CustomButton
         size="small"
@@ -40,7 +40,7 @@ const UseHint = () => {
         border="square"
         value="3개"
         onPress={() => setUseHint(3)}
-        // disabled={!isEnabled}
+      // disabled={!isEnabled}
       />
       <CustomButton
         size="small"
@@ -48,7 +48,7 @@ const UseHint = () => {
         border="square"
         value="4개 이상"
         onPress={() => setUseHint(4)} // 4 이상 사용했다고 가정
-        // disabled={!isEnabled}
+      // disabled={!isEnabled}
       />
     </View>
   );
